FEAT Standard energy seam calculations appear to be working.
There's a bug (of sorts) in Avix Das' anti-greedy test case: the
data works whether you're referencing the energy grid or the
seam calaculation grid.  Which means that it's possible to
confuse the two.

I only discovered this after I rotated the grid to try and
figure out how to calculate the horizontal seam.  I was
repeatedly getting the wrong data, and it turned out that
I was dereferencing the wrong array.  But now it's all
good.

Next up: seam an image!