FEAT Seam Carving!
Hey, look, seam carving works!  It works well, at least when removing
vertical seams.  I haven't tested out horizontal seam carving yet.
But this is the first step in getting it all working.

There are a number of obvious improvements: if we're doing multiple
carves, which we usually are, it probably makes sense to remove only the
column of seams that overlap or immediately neighbor the least-energy seam
from the energy map, and possibly from the dynamic map as well.  I don't
have to alloc/realloc the result image; just pare it down and then adjust
the height/width parameters manually (which sounds dangerous, natch).
And I could possibly multi-thread this, but that involves figuring out
how to break up the work into non-competing regions, which actually
isn't hard for the vertical seams, but sounds like a nightmare for the
horizontal ones, just due to the way memory is laid out.