FEAT ImageSeam is now a trait
The first step toward getting some control of this project is to define
a trait that describes how an image is turned into a seam.  By defining
it as two methods that take no arguments, seeding the seam list allows
us to eventually containerize our operations, giving the work a caching
and threading context.