FEAT Take an image, leave an image
After all that work, I'm finally back to the place where I started: I can
take an image and process it as Ixels, then convert it back into an image.

At least now I *have* Ixels, which means that I can also start to
process the image in-place.  The algorithms should be, well, if not
"straightforward" at least accessible.  I hope they're straightforward,
or this experiment has been one heck of a waste of time.